src/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --no-hashes
anyio==4.11.0
    # via starlette
blinker==1.9.0
    # via flask
cachetools==6.2.1
    # via google-auth
certifi==2025.10.5
    # via requests
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   flask
    #   functions-framework
    #   uvicorn
cloudevents==1.11.0
    # via functions-framework
colorama==0.4.6 ; sys_platform == 'win32'
    # via click
cryptography==46.0.3
    # via pyjwt
deprecation==2.1.0
    # via cloudevents
flask==3.1.2
    # via functions-framework
functions-framework==3.9.2
    # via gha-webhook
google-api-core==2.26.0
    # via
    #   google-cloud-appengine-logging
    #   google-cloud-compute-v1beta
    #   google-cloud-core
    #   google-cloud-logging
    #   google-cloud-secret-manager
google-auth==2.41.1
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-compute-v1beta
    #   google-cloud-core
    #   google-cloud-logging
    #   google-cloud-secret-manager
google-cloud-appengine-logging==1.7.0
    # via google-cloud-logging
google-cloud-audit-log==0.4.0
    # via google-cloud-logging
google-cloud-compute-v1beta==0.3.0
    # via gha-webhook
google-cloud-core==2.4.3
    # via google-cloud-logging
google-cloud-logging==3.12.1
    # via gha-webhook
google-cloud-secret-manager==2.25.0
    # via gha-webhook
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   google-cloud-audit-log
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.3
    # via
    #   google-cloud-logging
    #   google-cloud-secret-manager
grpcio==1.75.1
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-compute-v1beta
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.75.1
    # via google-api-core
gunicorn==23.0.0
    # via
    #   functions-framework
    #   uvicorn-worker
h11==0.16.0
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   requests
importlib-metadata==8.7.0
    # via opentelemetry-api
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
opentelemetry-api==1.38.0
    # via google-cloud-logging
packaging==25.0
    # via
    #   deprecation
    #   gunicorn
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-compute-v1beta
    #   google-cloud-logging
    #   google-cloud-secret-manager
protobuf==6.33.0
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-audit-log
    #   google-cloud-compute-v1beta
    #   google-cloud-logging
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.23 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pyjwt==2.10.1
    # via gha-webhook
requests==2.32.5
    # via google-api-core
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via anyio
starlette==0.48.0
    # via functions-framework
typing-extensions==4.15.0
    # via
    #   grpcio
    #   opentelemetry-api
urllib3==2.5.0
    # via requests
uvicorn==0.37.0
    # via
    #   functions-framework
    #   uvicorn-worker
uvicorn-worker==0.4.0
    # via functions-framework
watchdog==6.0.0
    # via functions-framework
werkzeug==3.1.3
    # via
    #   flask
    #   functions-framework
zipp==3.23.0
    # via importlib-metadata
